restructure rateengine pkg
